NN: merge r21573 from branches/jag/2.1-may12
git-svn-id: http://svn.zentyal.org/zentyal/trunk@21640 866bc4c1-54da-0310-a958-9d5a9a0e8d04

+++ main/core/src/EBox/Backup.pm
@@ -1124,7 +1124,7 @@ sub restoreBackup
         $self->_restoreEBoxEtcFiles($tempdir);
 
         # TODO: Make sure we don't open the file more than necessary
-        $self->_preRestoreActions($file);
+        $self->_preRestoreActions($file, %options);
 
         my @modules  = @{ $self->_modInstancesForRestore($file, %options) };
         my @restored = ();
@@ -1363,7 +1363,7 @@ sub _revokeRestore
 
 sub _preRestoreActions
 {
-    my ($self, $archive) = @_;
+    my ($self, $archive, %options) = @_;
 
     my $global = EBox::Global->getInstance();
     my @inBackup = @{ $self->_modulesInBackup($archive) };
@@ -1377,7 +1377,7 @@ sub _preRestoreActions
             push (@missing, $modName);
         }
     }
-    if (@missing) {
+    if (@missing and not $options{forceDependencies}) {
         throw EBox::Exceptions::External(
                 __x('The following modules present in the backup are not installed: {mods}. You need to install them before restoring.',
                     'mods' => join (' ', @missing))
